requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Django==2.2.2
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
toml==0.10.0
whitenoise==4.1.2
